chore: turned GeminiPrompt class into abstract class + added two implementation classes
